app/requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
bitarray==2.8.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cytoolz==0.12.2
dnspython==2.4.2
email-validator==2.0.0.post2
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.2
eth-utils==2.3.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastapi==0.103.1
frozenlist==1.4.0
geographiclib==2.0
geopy==2.4.0
h11==0.14.0
hexbytes==0.3.1
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
iso3166==2.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
loguru==0.7.2
lru-dict==1.2.0
MarkupSafe==2.1.3
multidict==6.0.4
nanoid==2.0.0
openpyxl==3.1.2
orjson==3.9.7
parsimonious==0.9.0
protobuf==4.25.1
pycryptodome==3.19.0
pydantic_core==2.6.3
pydantic==2.3.0
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
PyJWT==2.8.0
pymongo==4.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
pyunormalize==15.1.0
# pywin32==306
PyYAML==6.0.1
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
rlp==3.0.0
rpds-py==0.13.0
sniffio==1.3.0
starlette==0.27.0
toolz==0.12.0
typing_extensions==4.8.0
ujson==5.8.0
urllib3==2.1.0
uvicorn==0.23.2
# uvloop==0.17.0
watchfiles==0.20.0
web3==6.11.3
websockets==12.0
yarl==1.9.2
